requirements.txt
# LangChain 学习Demo - 依赖包列表
# 安装方式: pip install -r requirements.txt

# === 核心LangChain包 ===
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
python-dotenv>=1.0.0

# === 向量数据库 ===
chromadb>=1.0.0
# 注意: 如果需要FAISS，建议使用conda安装：
# conda install -c conda-forge faiss-cpu
# 或者尝试: pip install faiss-cpu

# === 文本处理 ===
tiktoken>=0.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# === Web应用和UI ===
streamlit>=1.40.0

# === 数据处理 ===
numpy>=2.0.0
pandas>=2.0.0

# === 已经通过依赖自动安装的包 ===
# pydantic>=2.0.0  # LangChain依赖
# openai>=1.0.0    # langchain-openai依赖
# aiohttp>=3.8.0   # langchain-community依赖